Refactoring and minor UI improvements - Move Kaltura-specific endpoints to own file - Move Kaltura-related helper functions to Kaltura files - By default, switch to the Kaltura player - Improve Tapestry settings UX - Some linting fixes
