+++ classes/class.constants.php
@@ -22,21 +22,6 @@ class CommentTypes
     public const STATUS_CHANGE = 'StatusChange';
 }
 
-class KalturaUpload
-{
-    // When updating STOP_UPLOAD_OPTION, autoload should be set to false
-    // so that we can clear the option from the cache when needed
-    public const STOP_UPLOAD_OPTION = 'tapestry_kaltura_stop_upload_requested';
-
-    public const IN_PROGRESS_OPTION = 'tapestry_kaltura_upload_in_progress';
-    public const LATEST_TAPESTRY_OPTION = 'tapestry_kaltura_upload_tapestry_id';
-    public const UPLOAD_LOG_OPTION = 'tapestry_kaltura_upload_log';
-    public const UPLOAD_ERROR_OPTION = 'tapestry_kaltura_upload_error';
-    public const YES_VALUE = 'yes';
-    public const NO_VALUE = 'no';
-    public const UPLOAD_BATCH_SIZE = 10;
-}
-
 class UploadStatus
 {
     public const NOT_STARTED = 'Not started';
