+++ classes/class.kaltura-api.php
@@ -31,9 +31,9 @@
         {
             $user = wp_get_current_user()->ID;
 
-            $kaltura_admin_secret = TapestryHelpers::getKalturaAdminSecret();
-            $kaltura_partner_id = TapestryHelpers::getKalturaPartnerId();
-            $kaltura_service_url = TapestryHelpers::getKalturaServiceUrl();
+            $kaltura_admin_secret = self::getKalturaAdminSecret();
+            $kaltura_partner_id = self::getKalturaPartnerId();
+            $kaltura_service_url = self::getKalturaServiceUrl();
 
             $kconf = new Configuration($kaltura_partner_id);
             $kconf->setServiceUrl($kaltura_service_url);
@@ -135,6 +135,10 @@
             }
         }
 
+        /*******************************************************/
+        /**                  PRIVATE FUNCTIONS                **/
+        /*******************************************************/
+
         private function _getOrCreateCategory($categoryName, $parentCategory, $categories, &$kAdminClient)
         {
             $categoryFullName = $parentCategory ? $parentCategory->fullName.'>'.$categoryName : $categoryName;
@@ -153,5 +157,112 @@
             }
 
             return $category;
+        }
+
+        /*******************************************************/
+        /**                 STATIC FUNCTIONS                  **/
+        /*******************************************************/
+
+        /**
+         * Getters for Kaltura configuration variables.
+         * Return null if LOAD_KALTURA is false.
+         */
+        public static function getKalturaAdminSecret()
+        {
+            return KALTURA_OVERRIDE_CONFIG ? get_option('kaltura_admin_secret') : (KALTURA_DEFAULT_CONFIG ? KALTURA_ADMIN_SECRET : null);
+        }
+
+        public static function getKalturaPartnerId()
+        {
+            return KALTURA_OVERRIDE_CONFIG ? get_option('kaltura_partner_id') : (KALTURA_DEFAULT_CONFIG ? KALTURA_PARTNER_ID : null);
+        }
+
+        public static function getKalturaServiceUrl()
+        {
+            return KALTURA_OVERRIDE_CONFIG ? get_option('kaltura_service_url') : (KALTURA_DEFAULT_CONFIG ? KALTURA_SERVICE_URL : null);
+        }
+
+        public static function getKalturaUniqueConfig()
+        {
+            return KALTURA_OVERRIDE_CONFIG ? get_option('kaltura_unique_config') : (KALTURA_DEFAULT_CONFIG ? KALTURA_UNIQUE_CONFIG : null);
+        }
+
+        /**
+         * Return the name of the Kaltura category a video should be sorted under.
+         */
+        public static function getKalturaCategoryName($tapestry)
+        {
+            if ('tapestry_name' === get_option('kaltura_category_structure')) {
+                return $tapestry->getSettings()->title;
+            } else {
+                // Categorize by date by default
+                return date('Y/m/d');
+            }
+        }
+
+        /**
+         * Update video node data and delete the local video after it has been uploaded to Kaltura.
+         *
+         * @param TapestryNode $node             Video node to update
+         * @param MediaEntry   $kalturaData      Response from Kaltura API
+         * @param bool         $useKalturaPlayer If true, also switch the video to use Kaltura player
+         * @param string       $videoPath        path to the video file
+         */
+        public static function saveAndDeleteLocalVideo($node, $kalturaData, $useKalturaPlayer, $videoPath)
+        {
+            $node->set((object) ['mediaFormat' => 'kaltura']);
+            $typeData = $node->getTypeData();
+            $typeData->mediaURL = $kalturaData->dataUrl.'?.mp4';
+            $typeData->kalturaId = $kalturaData->id;
+
+            // Save Kaltura account info so we can still show Kaltura player, even if LOAD_KALTURA is currently false
+            if (!isset($typeData->kalturaData)) {
+                $typeData->kalturaData = [];
+            }
+            $typeData->kalturaData['partnerId'] = self::getKalturaPartnerId();
+            $typeData->kalturaData['serviceUrl'] = self::getKalturaServiceUrl();
+            $typeData->kalturaData['uniqueConfiguration'] = self::getKalturaUniqueConfig();
+
+            $typeData->videoPlayer = $useKalturaPlayer ? 'kaltura' : 'regular';
+
+            $node->save();
+
+            wp_delete_file($videoPath);
+        }
+
+        /**
+         * Checks if the user has defined a maximum video upload size for Kaltura that is smaller than the WordPress max upload size,
+         * and if so, whether a video is too large to be uploaded.
+         *
+         * Assumes that the input video can be uploaded to Kaltura (check before calling).
+         *
+         * @param TapestryNode $node
+         *
+         * @return bool True if the user has defined no maximum video upload size, or it is not smaller than the WordPress max upload size.
+         *              Otherwise, returns true if the video is within the user-defined limit.
+         */
+        public static function checkVideoFileSize($node)
+        {
+            // Overridden value in WordPress settings
+            $max_file_size_setting = get_option('tapestry_kaltura_upload_max_file_size');
+
+            // Value defined in wp-config.php
+            $max_file_size_constant = defined('TAPESTRY_KALTURA_UPLOAD_MAX_FILE_SIZE') ? TAPESTRY_KALTURA_UPLOAD_MAX_FILE_SIZE : null;
+
+            $user_defined_size_string = !empty($max_file_size_setting) ? $max_file_size_setting : $max_file_size_constant;
+            if (empty($user_defined_size_string)) {
+                return true;
+            }
+
+            $user_defined_max_upload_size = wp_convert_hr_to_bytes($user_defined_size_string);
+
+            if ($user_defined_max_upload_size >= wp_max_upload_size()) {
+                return true;
+            }
+
+            $file = TapestryHelpers::getPathToNodeMedia($node);
+            $filesize = TapestryHelpers::getRealFileSize($file->file_path);
+
+            return $filesize <= $user_defined_max_upload_size;
         }
     }
